Module ChipExchange : pause task เขียน steps backend ของ MoneyToChips, MoneyToBonusChips, VoucherToBonusChips

+++ src/CasinovaAll/Modules/ChipExchange/TheS.Casinova.ChipExchange.Models/Models/ExchangeInformation.cs
@@ -25,7 +25,7 @@ namespace TheS.Casinova.ChipExchange.Models
         /// <summary>
         /// หมายเลขบัญชี
         /// </summary>
-        public int AccountNo { get; set; }
+        public double AccountNo { get; set; }
 
         /// <summary>
         /// รหัสตรวจสอบหมายเลขบัญชี
