+++ src/CasinovaAll/Modules/PlayerAccount/TheS.Casinova.PlayerAccount.Models/Models/PlayerAccountInformation.cs
@@ -43,7 +43,7 @@ namespace TheS.Casinova.PlayerAccount.Models
         /// <summary>
         /// หมายเลขบัญชี
         /// </summary>
-        public int AccountNo { get; set; }
+        public double AccountNo { get; set; }
 
         /// <summary>
         /// รหัสตรวจสอบหมายเลขบัญชี
